Move reading and writing helpers into dedicated file
