+++ Sources/ZIPFoundation/Archive+BackingConfiguration.swift
@@ -2,7 +2,10 @@
 //  Archive+BackingConfiguration.swift
 //  ZIPFoundation
 //
-//  Created by Thomas Zoechling on 19.09.21.
+//  Copyright © 2017-2021 Thomas Zoechling, https://www.peakstep.com and the ZIP Foundation project authors.
+//  Released under the MIT License.
+//
+//  See https://github.com/weichsel/ZIPFoundation/blob/master/LICENSE for license information.
 //
 
 import Foundation
